调整代码，controlnet独立
